Fix/crash on unavailable backend api (#199)
Staging Backend was unavailable it responded with html body (default nginx configuration) instead of json, which caused an exception. The number of exponential retries was set to 3 which meant the last try was done only after 7 seconds. I increased the number of retries, so last try will be after half an hour.